Change naming system on idioms
