+++ app/src/main/kotlin/com/jpaya/englishisfun/conditionals/data/network/model/ConditionalsResponse.kt
@@ -20,5 +20,5 @@ package com.jpaya.englishisfun.conditionals.data.network.model
  * Class to map the conditionals' FireStore response.
  */
 class ConditionalsResponse {
-    var conditionals: MutableList<ConditionalsNetworkItem> = mutableListOf()
+    var conditionals: MutableList<ConditionalNetworkItem> = mutableListOf()
 }
