+++ app/src/main/kotlin/com/jpaya/englishisfun/idioms/data/db/IdiomRoomItem.kt
@@ -14,16 +14,14 @@
  * limitations under the License.
  */
 
-package com.jpaya.englishisfun.conditionals.data.network.model
+package com.jpaya.englishisfun.idioms.data.db
 
-/**
- * Class to map the conditionals' FireStore response.
- */
-class ConditionalsNetworkItem {
-    var id: Long = 0
-    var name: String = ""
-    var condition: String = ""
-    var result: String = ""
-    var uses: MutableList<String> = mutableListOf()
-    var examples: MutableList<String> = mutableListOf()
-}
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity(tableName = "idioms")
+data class IdiomRoomItem(
+    @PrimaryKey(autoGenerate = true) val id: Long,
+    val idiom: String,
+    val description: String
+)
